Template/requirements.txt
browserstack-local==1.2.12
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
greenlet==3.0.3
grpcio==1.59.0
grpcio-tools==1.59.0
idna==3.10
natsort==8.4.0
openpyxl==3.1.5
overrides==7.7.0
playwright==1.44.0
protobuf==4.24.4
psutil==7.0.0
pyee==11.1.0
PyJWT==2.10.1
requests==2.32.4
robotframework==6.1.1
robotframework-assertion-engine==2.0.0
robotframework-browser==17.5.2
robotframework-pabot==4.2.0
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-stacktrace==0.4.1
setuptools==80.9.0
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
